project7

+++ src/board-project/pages/PostDetailPage/PostDetailPage.jsx
@@ -30,6 +30,42 @@ export default function PostDetailPage() {
   const [post, setPost] = useState(null); 
   const [loading , setLoading] = useState(false);
   const [error , setError] = useState("");
+  const [editCommentId , setEditCommentId] = useState(null);
+  const [editCommentInput , setEditCommentInput] = useState("");
+
+  // 댓글 수정.
+  const handleStartEditComment = comment => {
+    setEditCommentId(comment.commentId);
+    setEditCommentInput(comment.content ?? "");
+  }
+
+  // 댓글 수정 취소.
+  const handleCancelEditComment = () => {
+    setEditCommentId(null);
+    setEditCommentInput("");
+  }
+
+  // 댓글 수정 완료.
+  const handleUpdateComment = async commentId => {
+    const isLogin = !! localStorage.getItem("accessToken");
+    if(!isLogin) {
+      alert("로그인 후 댓글 수정이 가능합니다.");
+      navigate("/");
+      return;
+    }
+    const content = editCommentInput.trim();
+    if(!content) { alert("댓글 내용을 입력하세요."); return; }
+    const ok = window.confirm("댓글을 수정하시겠습니까?");
+    if(!ok) return;
+
+    try {
+      await sendJson(`/comments/${commentId}` , "PUT" , {content});
+      setEditCommentId(null);
+      setEditCommentInput("");
+      await loadComments();
+    } catch (e) { alert(e.message); }
+  };
+
 
   // 댓글 get요청. 에러 발생시 getJson에서 알아서 처리. 
   const loadComments = async () => {
@@ -37,7 +73,9 @@ export default function PostDetailPage() {
 
     // commentCount. deleted가 false인 애들은 제외시키고 count함.
     const aliveCount = data.filter( c => !c.deleted).length;
-    setPost( prev => prev ? ({...prev , comments : data , commentCount : aliveCount}) : prev);
+    setPost( prev => {if(!prev) { return prev; }
+  return{ ...prev , comments : data , commentCount : aliveCount};
+  })
   };
 
   // 좋아요는 자주 조작되므로 state로 따로 저장해서 관리. 
@@ -60,8 +98,10 @@ export default function PostDetailPage() {
       try { // 게시글 가져오기 실행.
         // ${pid}떄문에 ``(백틱) 써줘야함. "" 써주면 에러. --> 이거때문에 에러 생겼음.
         const data = await getJson(`/posts/${pid}`);
-        setPost(data); // 게시글 가져오기.
-        await loadComments(); // 그 게시글의 댓글 가져오기.
+        const commentData = await getJson(`/posts/${pid}/comments`);
+        const aliveCount = commentData.filter( c => !c.deleted).length;
+
+        setPost({...data , comments : commentData , commentCount : aliveCount});
         setLike(!!data.liked); // truthy(1,"1"...)는 true로 falsy(null,undefined,Nan...)는 false로 강제변환.
         setLikeCount(Number(data.likeCount) || 0); // Number(data.likeCount)가 숫자가 아니거나 이상한 값이면 0으로 안전처리. 
         // 0도 falsy인데 0을 0으로 처리? 이상하니까 전부 Number로 형변환. 
@@ -75,9 +115,15 @@ export default function PostDetailPage() {
 
   // post가 null이면 에러 만들지말고 그냥 undefined 반환해서 []로 다시 반환. , post가 객체가 있으면 comments 반환. 
   // 의존성을 post로 한 이유 : 게시글이 댓글도 바뀔 수 있다. -> post가 바뀌면 commentAll함수 재실행. 
-  const commentList = useMemo( () => commentAll(post?.comments || []) , [post]);
+  const commentList = useMemo( () => commentAll(post?.comments || []) , [post?.comments]);
 
   const handleClickLike = async () => { // 좋아요 클릭했을때.
+    const isLogin = !!localStorage.getItem("accessToken");
+    if(!isLogin) {
+      alert("로그인 후 좋아요 가능합니다.");
+      navigate("/login");
+      return;
+    }
     try {
       const method = like ? "DELETE" : "POST"; // like가 true인 상태에서 클릭했으면 false로 변환되니까 DELETE 보냄.
 
@@ -91,8 +137,18 @@ export default function PostDetailPage() {
 
   const handleAddComment = async (e) => { // 댓글 등록할때.
     e.preventDefault(); // form submit하면 브라우저는 새로고침하지만 그걸 막는 코드.
+    const isLogin = !!localStorage.getItem("accessToken");
     const content = commentInput.trim(); // 공백만 있는 댓글은 등록 안하는 규칙.
+
     if (!content) return; // content가 빈문자열(false)이면 무시.(return)
+    if(!isLogin) {
+      alert("로그인 후 댓글 작성이 가능합니다.");
+      navigate("/login");
+      return;
+    }
+
+    const ok = window.confirm("댓글을 등록하시겠습니까?");
+    if(!ok) return;
 
     try {
       await sendJson(`/posts/${pid}/comments` , "POST" , {
@@ -105,8 +161,19 @@ export default function PostDetailPage() {
 
   const handleAddReply = async (e) => {
     e.preventDefault();
+
+    const isLogin = !!localStorage.getItem("accessToken");
     const content = replyInput.trim();
+
     if (!content || replyTo == null) return; // 입력한 대댓글이 빈문자열 || 대댓글 대상(replyTo)이 없으면 무시.(return)
+    if(!isLogin) {
+      alert("로그인 후 답글 작성이 가능합니다.");
+      navigate("/login");
+      return;
+    }
+
+    const ok = window.confirm("답글을 등록하시겠습니까?");
+    if(!ok) return;
 
     try {
       await sendJson(`/posts/${pid}/comments` , "POST" , {
@@ -120,6 +187,9 @@ export default function PostDetailPage() {
   };
 
   const handleDeleteComment = async (commentId) => { // 댓글 삭제.
+    const ok = window.confirm("이 댓글을 삭제하시겠습니까?");
+    if(!ok) return;
+    
     try {
        await deleteJson(`/comments/${commentId}`);
        await loadComments(); // 최신 댓글목록으로 업데이트.
@@ -143,6 +213,9 @@ export default function PostDetailPage() {
   if (error) return <div className="postPage"><div className="emptyBox">에러: {error}</div></div>;
   if (!post) return <div className="postPage"><div className="emptyBox">게시글이 없습니다.</div></div>;
 
+  const loginUserId = Number(localStorage.getItem("loginUserId")); // 로컬스토리지에 있는 userId꺼내서 현재 user랑 같은지 비교.
+  const isMyPost = loginUserId == Number(post.userId);
+
   return (
     <div className="postPage">
       <div className="topButtons">
@@ -150,14 +223,18 @@ export default function PostDetailPage() {
           목록으로
         </button>
 
+        {isMyPost && (
         <button className="editBtn" type="button" onClick={ () => navigate(`/posts/${pid}/edit`)}>
           수정하기
         </button>
-      </div>
+        )}
+        </div>
 
-      <button className="deleteBtn" onClick={handleDeletePost}>
+      {isMyPost && (
+        <button className="deleteBtn" onClick={handleDeletePost}>
         삭제하기
       </button>
+      )}
 
       <header className="postHeader">
         <h1 className="postTitle">{post.title}</h1>
@@ -202,13 +279,13 @@ export default function PostDetailPage() {
       <section className="attachSection">
         <div className="sectionTitle">첨부파일</div>
 
-        {post.attachments.length === 0 ? (
+        {(post.attachments ?? []).length === 0 ? (
           <div className="emptyBox">첨부파일이 없습니다.</div>
         ) 
         : 
         (
           <ul className="attachList">
-            {[...post.attachments]
+            {[...(post.attachments ?? [])]
               .sort( (a , b) => (a.sortOrder ?? 0) - (b.sortOrder ?? 0))
               .map( f => (
                 <li key={f.id} className="attachItem">
@@ -258,6 +335,13 @@ export default function PostDetailPage() {
                 setReplyInput={setReplyInput}
                 handleAddReply={handleAddReply}
                 handleDeleteComment={handleDeleteComment}
+                loginUserId={loginUserId}
+                editCommentId={editCommentId}
+                editCommentInput={editCommentInput}
+                setEditCommentInput={setEditCommentInput}
+                handleStartEditComment={handleStartEditComment}
+                handleCancelEditComment={handleCancelEditComment}
+                handleUpdateComment={handleUpdateComment}
               />
             ))
           )}
@@ -268,9 +352,12 @@ export default function PostDetailPage() {
 }
 
 function CommentItem({
-  comment , depth , replyTo , setReplyTo , replyInput , setReplyInput , handleAddReply , handleDeleteComment
+  comment , depth , replyTo , setReplyTo , replyInput , setReplyInput , handleAddReply , handleDeleteComment , loginUserId ,
+  editCommentId , editCommentInput , setEditCommentInput , handleStartEditComment: handleEditComment , handleCancelEditComment , handleUpdateComment
 }) {
   const isReplying = replyTo === comment.commentId;
+  const isMyComment = Number(loginUserId) === Number(comment.userId);
+  const isEditing = editCommentId === comment.commentId;
 
   return (
     <div className={`commentItem depth-${depth}`}>
@@ -290,19 +377,43 @@ function CommentItem({
             >
               답글
             </button>
-            <button
+            {isMyComment && !comment.deleted && (
+              <button type="button" className="linkBtn" onClick={ () => handleEditComment(comment)}>
+                수정
+              </button>
+            )}
+            {isMyComment && (
+              <button
               type="button"
               className="linkBtn danger"
               onClick={ () => handleDeleteComment(comment.commentId)}
             >
               삭제
             </button>
+            )}
           </div>
         </div>
 
-        <div className={`commentContent ${comment.deleted ? "deleted" : ""}`}>
-          {comment.deleted ? "삭제된 댓글입니다." : comment.content}
-        </div>
+        {isEditing ? (
+          <div className="editCommentBox">
+            <textarea className="replyInput" value={editCommentInput} onChange={ e => setEditCommentInput(e.target.value)}
+              placeholder="댓글을 수정하세요." />
+            <div className="replyBtns">
+              <button type="button" className="replySubmit" onClick={ () => handleUpdateComment(comment.commentId)}>
+                수정 완료
+              </button>
+              <button type="button" className="replyCancel" onClick={handleCancelEditComment}>
+                수정 취소
+              </button>
+            </div>    
+          </div> )
+          :
+          (
+            <div className={`commentContent ${comment.deleted ? "deleted" : ""}`}>
+              {comment.deleted ? "삭제된 댓글입니다." : comment.content}
+            </div>
+          )
+        }
 
         {isReplying && (
           <form className="replyForm" onSubmit={handleAddReply}>
@@ -337,6 +448,13 @@ function CommentItem({
               setReplyInput={setReplyInput}
               handleAddReply={handleAddReply}
               handleDeleteComment={handleDeleteComment}
+              loginUserId={loginUserId}
+              editCommentId={editCommentId}
+              editCommentInput={editCommentInput}
+              setEditCommentInput={setEditCommentInput}
+              handleStartEditComment={handleEditComment}
+              handleCancelEditComment={handleCancelEditComment}
+              handleUpdateComment={handleUpdateComment}
             />
           ))}
         </div>
